requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.43.3
bcbio-gff==0.7.1
bio==1.5.9
biopython==1.81
biothings-client==0.3.1
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
Eel==0.16.0
et-xmlfile==1.1.0
fonttools==4.51.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
gprofiler-official==1.0.0
greenlet==3.0.3
idna==3.7
kiwisolver==1.4.5
matplotlib==3.8.4
mygene==3.2.2
networkx==3.3
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
primer3-py==2.0.1
pycparser==2.22
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
requests==2.31.0
six==1.16.0
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.3
